+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # `nonempty-collections`
 
+## Unreleased
+
+#### Added
+
+- vector: `NEVec::reverse`, mysteriously missing until this point.
+
+## 1.3.0 (2026-01-13)
+
+#### Added
+
+- vector: `NEVec::from_elem` exposed for times when you want `nev![foo; n]` but `n` is not `const`.
+- slice: `NESlice::last` to match `NEVec`.
+
+#### Changed
+
+- vector: The `Index` impl has been generalized to accept more slice patterns.
+
 ## 1.2.1 (2025-11-11)
 
 Happy Remembrance Day.